% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internet.R
\name{CLAP}
\alias{CLAP}
\title{Clap between words}
\usage{
CLAP(x, emoji = "👏")
}
\arguments{
\item{x}{A character vector.}

\item{emoji}{The emoji to insert (default clap).}
}
\value{
A character vector with emoji between words.
}
\description{
Add clap emoji between words for emphasis.
}
\examples{
CLAP("say it louder")
CLAP("do it now", emoji = "\U0001F525")
}
