% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotion.R
\name{DRAMATICALLY}
\alias{DRAMATICALLY}
\title{Speak dramatically}
\usage{
DRAMATICALLY(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector with dramatic pauses.
}
\description{
Add dramatic pauses between words.
}
\examples{
DRAMATICALLY("I cant believe it")
DRAMATICALLY("no this cant be")
}
