% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sound.R
\name{ECHO}
\alias{ECHO}
\title{Echo text}
\usage{
ECHO(x, times = 3)
}
\arguments{
\item{x}{A character vector.}

\item{times}{Number of echoes (default 3).}
}
\value{
A character vector with echo effect.
}
\description{
Repeat text with fading effect.
}
\examples{
ECHO("hello")
ECHO("goodbye", times = 4)
}
