% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sound.R
\name{ELONGATE}
\alias{ELONGATE}
\title{Elongate text}
\usage{
ELONGATE(x, n = 3)
}
\arguments{
\item{x}{A character vector.}

\item{n}{Number of times to repeat vowels (default 3).}
}
\value{
A character vector with elongated vowels.
}
\description{
Stretch vowels for emphasis.
}
\examples{
ELONGATE("no")
ELONGATE("why", n = 5)
}
