% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{INDOOR_VOICE}
\alias{INDOOR_VOICE}
\title{Use your indoor voice}
\usage{
INDOOR_VOICE(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector with a gentle reminder.
}
\description{
Politely remind to use lowercase.
}
\examples{
INDOOR_VOICE("HELLO")
}
