% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{IS_YELLING}
\alias{IS_YELLING}
\title{Check if text is yelling}
\usage{
IS_YELLING(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A logical vector.
}
\description{
Returns TRUE if text is all uppercase.
}
\examples{
IS_YELLING("HELLO")
IS_YELLING("hello")
IS_YELLING("Hello")
}
