% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internet.R
\name{MOCK}
\alias{MOCK}
\title{Mock text}
\usage{
MOCK(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector in mocking case.
}
\description{
SpongeBob mocking meme text with alternating case.
}
\examples{
MOCK("I'm not mad")
MOCK("this is fine")
}
