% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotion.R
\name{PASSIVE_AGGRESSIVELY}
\alias{PASSIVE_AGGRESSIVELY}
\title{Speak passive-aggressively}
\usage{
PASSIVE_AGGRESSIVELY(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector with periods between words.
}
\description{
Add periods between words.
}
\examples{
PASSIVE_AGGRESSIVELY("fine whatever sure")
PASSIVE_AGGRESSIVELY("I said its fine")
}
