% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotion.R
\name{SARCASTICALLY}
\alias{SARCASTICALLY}
\title{Speak sarcastically}
\usage{
SARCASTICALLY(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector in alternating case.
}
\description{
Alternating caps for sarcastic tone.
}
\examples{
SARCASTICALLY("oh wow thats great")
SARCASTICALLY("sure jan")
}
