% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{SHOUT_INTO_VOID}
\alias{SHOUT_INTO_VOID}
\title{Shout into the void}
\usage{
SHOUT_INTO_VOID(x)
}
\arguments{
\item{x}{A character vector (ignored).}
}
\value{
An empty string.
}
\description{
The void does not answer back.
}
\examples{
SHOUT_INTO_VOID("HELLO?")
SHOUT_INTO_VOID("IS ANYONE THERE?")
}
