% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sound.R
\name{STUTTER}
\alias{STUTTER}
\title{Stutter text}
\usage{
STUTTER(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector with stuttering.
}
\description{
Add stuttering effect to words.
}
\examples{
STUTTER("what do you mean")
STUTTER("I dont know")
}
