% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{WHISPER}
\alias{WHISPER}
\title{Whisper text}
\usage{
WHISPER(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector in lowercase.
}
\description{
Convert text to lowercase for whispering.
}
\examples{
WHISPER("PLEASE BE QUIET")
}
