% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YELL.R
\name{YELL}
\alias{YELL}
\title{YELL}
\usage{
YELL(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector with all characters converted to upper case.
}
\description{
YELL
}
\examples{
YELL("HELLO WORLD")
}
