% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{YELL_LOUDER}
\alias{YELL_LOUDER}
\title{Yell louder}
\usage{
YELL_LOUDER(x, n = 3)
}
\arguments{
\item{x}{A character vector.}

\item{n}{Number of exclamation marks to add (default 3).}
}
\value{
A character vector with added exclamation marks.
}
\description{
Add more exclamation marks to make text louder.
}
\examples{
YELL_LOUDER("HELLO!")
}
